requirements.txt
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
torchvision>=0.16.0+cu118
torch>=2.0.1+cu118

# pip install git+https://github.com/openai/CLIP.git@main
git+https://github.com/openai/CLIP.git@main

numpy>=1.24.1
open_clip_torch>=2.23.0
Pillow>=9.3.0
onnx>=1.14.1
